social_service/requirements.txt
# requirements.txt
Django==5.1.6
djangorestframework